requirements.txt
aws-cdk.aws-lambda
aws-cdk.aws-iam
aws-cdk.aws-dynamodb
cdk-dynamo-table-viewer